+++ allocator/elpekenin/allocator.h
@@ -163,7 +163,7 @@ size_t get_used_heap(void);
  *
  * :c:var:`n` will be set to the number of allocators.
  */
-const allocator_t **get_known_allocators(uint8_t *n);
+const allocator_t **get_known_allocators(size_t *n);
 
 /**
  * Get the allocator defined as "default".
