+++ indicators/introspection.c
@@ -7,11 +7,11 @@
 
 #define NUM_INDICATORS_RAW (sizeof(indicators) / sizeof(indicator_t))
 
-uint8_t indicators_count(void) {
+size_t indicators_count(void) {
     return NUM_INDICATORS_RAW;
 }
 
-indicator_t get_indicator(uint8_t index) {
+indicator_t get_indicator(size_t index) {
     indicator_t value;
     memcpy_P(&value, &indicators[index], sizeof(indicator_t));
     return value;
